Replace room and pos vars with GetRoom/SetRoom and GetPosition/SetPosition. Documentation cleanup, too.

+++ Common/Game/collision/floordata.h
@@ -3,7 +3,7 @@
 #include "Specific/trmath.h"
 #include "Specific/newtypes.h"
 
-constexpr auto WALL_PLANE = Vector3(0, 0, -CLICK(127));
+constexpr auto WALL_PLANE = Vector3(0, 0, static_cast<float>(-CLICK(127)));
 
 enum class FLOOR_MATERIAL : unsigned char
 {
